refactor(los): extract computed view properties into View structs
Extract 10 computed view properties from LineOfSightView.swift into
separate View structs, reducing it from 1335 to 711 lines (~47%).
Also collapse analysisSheetContent/analysisSheetVStack into a single
property. Pure structural refactor with no behavior changes.
